Add participant table and methods
Making sure that everything is consistent when the bot does not know the list of participants is
getting increasingly difficult. Plus, if the bot is used with large groups it may be difficult for
admins to make sure that participants are not listed under different names.

All in all, the cost of manually registering a participant list is low.

In light of this, this commit adds support in the database for an explicit list of participants.