Make all names lowercase in the parser layer
Participant names were case-insensitive from the beginning, but we used to save them in the database
using the case that the user provided, while the bot logic made everything lowercase. This was done
to leave open the possibility to later be able to print participants with the correct case (e.g.,
`Name` instead of `name`), without any breaking change to the database.

However, now that we are introducing groups and maybe later even aliases, we'll need to compare
names more often, and having to convert things to lowercase in multiple places is
inconvenient.

Also, printing names with the correct case is definitely not so crucial. If needed, it should be
supported by the \addparticipants command, and a breaking change in the database is acceptable.

Therefore, with this commit we turn everything into lowercase in the parser, so later we don't need
to care about it anymore.